--- runtime/src/ext/fetch/fetch/mod.rs
@@ -1,19 +1,0 @@
-// This Source Code Form is subject to the terms of the Mozilla Public
-// License, v. 2.0. If a copy of the MPL was not distributed with this
-// file, You can obtain one at https://mozilla.org/MPL/2.0/.
-
-use andromeda_core::Extension;
-
-#[derive(Default)]
-pub struct FetchExt;
-
-impl FetchExt {
-    pub fn new_extension() -> Extension {
-        Extension {
-            name: "fetch",
-            ops: vec![],
-            storage: None,
-            files: vec![include_str!("./mod.ts")],
-        }
-    }
-}
